depth values quantitized to byte

+++ DEPTH/Assets/Scripts/MainBehavior.cs
@@ -272,7 +272,7 @@ public class MainBehavior : MonoBehaviour {
 		Texture texture = _vp.texture;
 		if (texture == null) return;
 
-		float[] depths = null;
+		byte[] depths = null;
 		
 		//If depth file exists, try to read from it
 		if (_depthFilePath != null)
@@ -489,7 +489,7 @@ public class MainBehavior : MonoBehaviour {
 		_startFrame = 0;
 
 		int modelTypeVal = -1;
-		float[] depths;
+		byte[] depths;
 
 		//Check if the file was processed
 		if (_searchCache)
@@ -649,7 +649,7 @@ public class MainBehavior : MonoBehaviour {
 			_orig_height = texture.height;
 			_framecount = 1;
 
-			float[] depths = DepthFileUtils.ReadFromArchive(0, out _x, out _y);
+			byte[] depths = DepthFileUtils.ReadFromArchive(0, out _x, out _y);
 			_meshBehav.SetScene(depths, _x, _y, (float) _orig_width/_orig_height, texture);
 
 			DepthFilePanel.SetActive(true);
@@ -791,7 +791,7 @@ public class MainBehavior : MonoBehaviour {
 
 		if (_donnx == null) return;
 
-		float[] depths = _donnx.Run(texture, out _x, out _y);
+		byte[] depths = _donnx.Run(texture, out _x, out _y);
 		_meshBehav.SetScene(depths, _x, _y, (float) _orig_width/_orig_height, texture);
 	}
